refactor: simplify Foundry coverage gate
