+++ .github/workflows/ci.yml
@@ -99,50 +99,8 @@ jobs:
       - name: Deploy production scripts to localhost
         run: corepack yarn deploy:localhost
 
-  coverage-producers:
-    name: Coverage producer (${{ matrix.group }})
-    runs-on: ubuntu-latest
-    strategy:
-      fail-fast: false
-      matrix:
-        group:
-          - deterministic
-          - parity-main
-          - parity-collisions
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@11d5960a326750d5838078e36cf38b85af677262
-        with:
-          submodules: recursive
-      - name: Install pinned Foundry
-        uses: foundry-rs/foundry-toolchain@b00af27efadbc7b4ca8b82abbd903b17cc874d2a
-        with:
-          version: ${{ env.FOUNDRY_VERSION }}
-      - name: Set up pinned Node.js
-        uses: actions/setup-node@49933ea5288caeca8642d1e84afbd3f7d6820020
-        with:
-          node-version: ${{ env.NODE_VERSION }}
-          cache: yarn
-      - name: Install dependencies reproducibly
-        run: |
-          corepack enable
-          corepack yarn install --immutable
-      - name: Prepare local environment
-        run: cp .env.default .env
-      - name: Produce ${{ matrix.group }} coverage artifacts
-        run: node foundry-migration/tools/run-foundry-coverage.cjs --group "${{ matrix.group }}"
-      - name: Upload ${{ matrix.group }} coverage artifacts
-        uses: actions/upload-artifact@ea165f8d65b6e75b540449e92b4886f43607fa02
-        with:
-          name: coverage-${{ matrix.group }}
-          path: coverage
-          if-no-files-found: error
-          retention-days: 1
-
   coverage:
     name: Foundry coverage and Codecov
-    needs:
-      - coverage-producers
     runs-on: ubuntu-latest
     permissions:
       contents: read
@@ -156,14 +114,15 @@ jobs:
         uses: actions/setup-node@49933ea5288caeca8642d1e84afbd3f7d6820020
         with:
           node-version: ${{ env.NODE_VERSION }}
-      - name: Download parallel coverage artifacts
-        uses: actions/download-artifact@d3f86a106a0bac45b974a628896c90dbdf5c8093
-        with:
-          pattern: coverage-*
-          path: coverage
-          merge-multiple: true
-      - name: Merge coverage and enforce same-denominator gates
-        run: node foundry-migration/tools/run-foundry-coverage.cjs --merge-only
+          cache: yarn
+      - name: Install dependencies reproducibly
+        run: |
+          corepack enable
+          corepack yarn install --immutable
+      - name: Prepare local environment
+        run: cp .env.default .env
+      - name: Run deterministic Foundry coverage and enforce baseline floors
+        run: corepack yarn coverage
       - name: Upload Foundry LCOV to Codecov with OIDC
         uses: codecov/codecov-action@0fb7174895f61a3b6b78fc075e0cd60383518dac
         with:
