+++ README.md
@@ -525,10 +525,9 @@ See [TESTING.md](./TESTING.md) for suite design, seed reproduction, coverage mec
 
 ### Coverage
 
-- `corepack yarn coverage`: run the complete suite, build accurate Foundry LCOV, and enforce the same-denominator starting-behavior gates
-- `corepack yarn coverage:merge`: rebuild the merged report from previously generated shards after diagnosing a coverage run
+- `corepack yarn coverage`: run deterministic coverage once, build accurate Foundry LCOV, and enforce the permanent Foundry baseline floors
 
-Coverage is intentionally heavy and runs on every push and pull request. The output is `coverage/lcov.info`; CI uploads it to the `foundry` Codecov flag through OIDC and fails visibly on upload errors.
+Coverage runs on every pull request and every push to `main`. Fuzz and invariant tests remain outside coverage but always run in the complete-suite job. The output is `coverage/lcov.info`; CI uploads it to the `foundry` Codecov flag through OIDC and fails visibly on upload errors.
 
 ### Packaging Commands
