--- foundry-migration/baseline/hardhat-coverage-by-file.csv
@@ -1,36 +1,0 @@
-file,statements_covered,statements_total,statements_pct,branches_covered,branches_total,branches_pct,functions_covered,functions_total,functions_pct,lines_covered,lines_total,lines_pct
-contracts/Escrow.sol,196,196,100,220,220,100,53,53,100,268,268,100
-contracts/EscrowV2.sol,299,299,100,336,336,100,70,70,100,392,392,100
-contracts/Orchestrator.sol,107,107,100,96,96,100,24,24,100,136,136,100
-contracts/OrchestratorV2.sol,151,151,100,108,108,100,33,33,100,197,197,100
-contracts/OrchestratorV3.sol,38,38,100,24,24,100,15,15,100,50,50,100
-contracts/ProtocolViewer.sol,19,19,100,4,4,100,6,6,100,31,31,100
-contracts/ProtocolViewerV2.sol,26,26,100,4,4,100,8,8,100,41,41,100
-contracts/RateManagerV1.sol,50,50,100,60,60,100,15,15,100,70,70,100
-contracts/RiskManager.sol,201,208,96.63,174,246,70.73,39,40,97.5,273,295,92.54
-contracts/StakeVault.sol,200,216,92.59,135,226,59.73,44,49,89.8,287,316,90.82
-contracts/hooks/AcrossBridgeHook.sol,29,29,100,28,28,100,7,7,100,36,36,100
-contracts/hooks/AcrossBridgeHookV2.sol,31,31,100,30,30,100,7,7,100,39,39,100
-contracts/hooks/SignatureGatingPreIntentHook.sol,15,15,100,14,14,100,4,4,100,21,21,100
-contracts/hooks/WhitelistPreIntentHook.sol,19,19,100,20,20,100,6,6,100,25,25,100
-contracts/lib/BoundedCall.sol,18,19,94.74,15,18,83.33,4,4,100,21,22,95.45
-contracts/lib/FeeSettlementLib.sol,26,26,100,14,14,100,3,3,100,34,34,100
-contracts/lib/PostIntentHookExecutor.sol,13,13,100,4,6,66.67,2,2,100,13,13,100
-contracts/lib/ReferralFeeLib.sol,11,11,100,10,10,100,2,2,100,17,17,100
-contracts/lib/RiskSettlementExecutor.sol,15,15,100,8,8,100,1,1,100,18,18,100
-contracts/lib/ThresholdSigVerifierUtils.sol,14,14,100,16,16,100,1,1,100,21,21,100
-contracts/oracles/ChainlinkOracleAdapter.sol,26,26,100,18,18,100,3,3,100,30,30,100
-contracts/oracles/PythOracleAdapter.sol,23,23,100,12,12,100,4,4,100,29,29,100
-contracts/registries/EscrowRegistry.sol,10,10,100,12,12,100,6,6,100,13,13,100
-contracts/registries/NullifierRegistry.sol,10,10,100,14,14,100,6,6,100,14,14,100
-contracts/registries/NullifierRegistryV2.sol,17,17,100,26,26,100,7,7,100,24,24,100
-contracts/registries/OrchestratorRegistry.sol,6,6,100,12,12,100,2,2,100,8,8,100
-contracts/registries/PaymentVerifierRegistry.sol,31,31,100,28,28,100,12,12,100,36,36,100
-contracts/registries/PostIntentHookRegistry.sol,9,9,100,10,10,100,5,5,100,11,11,100
-contracts/registries/RelayerRegistry.sol,8,8,100,10,10,100,4,4,100,10,10,100
-contracts/unifiedVerifier/BaseUnifiedPaymentVerifier.sol,14,14,100,18,18,100,7,7,100,21,21,100
-contracts/unifiedVerifier/BaseUnifiedPaymentVerifierV3.sol,17,17,100,24,24,100,7,7,100,24,24,100
-contracts/unifiedVerifier/MultiAttestationVerifier.sol,21,21,100,24,26,92.31,8,8,100,24,24,100
-contracts/unifiedVerifier/SimpleAttestationVerifier.sol,5,5,100,4,4,100,3,3,100,9,9,100
-contracts/unifiedVerifier/UnifiedPaymentVerifier.sol,35,35,100,28,28,100,11,11,100,39,39,100
-contracts/unifiedVerifier/UnifiedPaymentVerifierV3.sol,41,41,100,40,40,100,11,11,100,45,45,100
