--- foundry-migration/coverage-exceptions.json
@@ -1,55 +1,0 @@
-{
-  "schemaVersion": 1,
-  "exceptions": [
-    {
-      "file": "contracts/Orchestrator.sol",
-      "metric": "statements",
-      "reason": "Solidity 0.8.18 reports two statement anchors without stable cross-run identities when the ten exact-source legacy test shards are merged. The same merged report covers 160/160 lines, 39/39 branches, and 24/24 functions, and all legacy behaviors have deterministic parity mappings."
-    },
-    {
-      "file": "contracts/OrchestratorV2.sol",
-      "metric": "lines",
-      "reason": "Standard non-IR coverage cannot compile this stack-heavy production contract; the required minimal-IR fallback warns that three executed source anchors cannot be mapped. Production refactoring solely for instrumentation is prohibited, while deterministic parity covers every V2 behavior."
-    },
-    {
-      "file": "contracts/OrchestratorV2.sol",
-      "metric": "statements",
-      "reason": "Standard non-IR coverage cannot compile this stack-heavy production contract; minimal-IR maps 250/252 of its expanded statement denominator and reports unmappable executed anchors. Deterministic parity and the line/branch/function report provide compensating evidence without changing production code."
-    },
-    {
-      "file": "contracts/OrchestratorV2.sol",
-      "metric": "branches",
-      "reason": "The minimal-IR fallback required for this stack-heavy contract loses five branch anchors even though the corresponding deterministic success and revert scenarios execute. The compiler emits explicit anchor warnings; production code is intentionally unchanged."
-    },
-    {
-      "file": "contracts/OrchestratorV3.sol",
-      "metric": "lines",
-      "reason": "After upstream f6fd7aa expanded standalone V3, Solidity 0.8.18 minimal-IR leaves only the pause/unpause wrapper lines unanchored even though exact V3 governance tests execute both and all 35 functions plus all 39 branches are covered. Standard coverage cannot compile FeeSettlementLib because of stack-too-deep, so production is not refactored solely for instrumentation."
-    },
-    {
-      "file": "contracts/OrchestratorV3.sol",
-      "metric": "statements",
-      "reason": "The only two uncovered minimal-IR statement anchors are the executed pause/unpause wrappers. Exact V3 owner/non-owner governance tests pass, function and branch coverage are 100%, and standard coverage fails stack-too-deep in FeeSettlementLib before tests can run."
-    },
-    {
-      "file": "contracts/RateManagerV1.sol",
-      "metric": "statements",
-      "reason": "Standard coverage fails stack-too-deep in RateManagerV1, while minimal-IR reports one unanchored statement in an exercised path. The report still covers 84/84 lines and 15/15 functions, and all rate-manager parity scenarios execute."
-    },
-    {
-      "file": "contracts/RateManagerV1.sol",
-      "metric": "branches",
-      "reason": "Standard coverage fails stack-too-deep in RateManagerV1; minimal-IR loses one branch anchor and reports 22/23. Deterministic tests cover configuration, authorization, fixed/oracle fallback, and update paths without a production-only instrumentation refactor."
-    },
-    {
-      "file": "contracts/RiskManager.sol",
-      "metric": "statements",
-      "reason": "Foundry expands RiskManager's statement denominator from Hardhat's 208 to 432 and standard coverage cannot compile the stack-heavy contract. Foundry nevertheless covers 407 statements versus 201 at baseline, while lines improve 92.54% to 96.66%, branches improve 70.73% to 71.74%, and functions improve 97.50% to 100%."
-    },
-    {
-      "file": "contracts/lib/ThresholdSigVerifierUtils.sol",
-      "metric": "statements",
-      "reason": "Both standard and minimal-IR coverage report one unanchored assembly-adjacent statement even though the path executes. The exact-source shard covers 22/22 lines, 12/12 branches, and the sole function, with dedicated threshold/signature tests."
-    }
-  ]
-}
