+++ package.json
@@ -19,8 +19,7 @@
     "chain": "hardhat node --no-deploy",
     "clean": "rm -rf cache cache_forge coverage artifacts out typechain dist",
     "compile": "hardhat compile",
-    "coverage": "node foundry-migration/tools/run-foundry-coverage.cjs",
-    "coverage:merge": "node foundry-migration/tools/run-foundry-coverage.cjs --merge-only",
+    "coverage": "node scripts/run-foundry-coverage.cjs",
     "deploy:localhost": "ts-node --transpile-only scripts/deployActive.ts localhost",
     "deploy:sepolia": "ts-node --transpile-only scripts/deployActive.ts sepolia",
     "deploy:base": "ts-node --transpile-only scripts/deployActive.ts base && hardhat export --network base --export ./deployments/outputs/baseContracts.ts && yarn workspace @zkp2p/contracts-v2 extract",
