+++ test-foundry/coverage-exceptions.json
@@ -1,0 +1,55 @@
+{
+  "schemaVersion": 1,
+  "exceptions": [
+    {
+      "file": "contracts/Orchestrator.sol",
+      "metric": "statements",
+      "reason": "Solidity 0.8.18 reports two statement anchors without stable cross-run identities when the ten exact-source legacy test shards are merged. The merged report still covers every line, branch, and function in this contract."
+    },
+    {
+      "file": "contracts/OrchestratorV2.sol",
+      "metric": "lines",
+      "reason": "Standard non-IR coverage cannot compile this stack-heavy production contract; the required minimal-IR fallback warns that three executed source anchors cannot be mapped. Production code is not refactored solely for instrumentation."
+    },
+    {
+      "file": "contracts/OrchestratorV2.sol",
+      "metric": "statements",
+      "reason": "Standard non-IR coverage cannot compile this stack-heavy production contract; minimal-IR maps 250 of 252 expanded statement anchors and reports the remaining executed anchors as unmappable."
+    },
+    {
+      "file": "contracts/OrchestratorV2.sol",
+      "metric": "branches",
+      "reason": "The minimal-IR fallback required for this stack-heavy contract loses five branch anchors even though the corresponding deterministic success and revert scenarios execute."
+    },
+    {
+      "file": "contracts/OrchestratorV3.sol",
+      "metric": "lines",
+      "reason": "Solidity 0.8.18 minimal-IR leaves only the exercised pause and unpause wrapper lines unanchored. Standard coverage cannot compile FeeSettlementLib because of stack-too-deep."
+    },
+    {
+      "file": "contracts/OrchestratorV3.sol",
+      "metric": "statements",
+      "reason": "The only two uncovered minimal-IR statement anchors are the exercised pause and unpause wrappers. Function and branch coverage are complete, while standard coverage fails stack-too-deep in FeeSettlementLib."
+    },
+    {
+      "file": "contracts/RateManagerV1.sol",
+      "metric": "statements",
+      "reason": "Standard coverage fails stack-too-deep in RateManagerV1, while minimal-IR reports one unanchored statement in an exercised path. The report still covers every line and function."
+    },
+    {
+      "file": "contracts/RateManagerV1.sol",
+      "metric": "branches",
+      "reason": "Standard coverage fails stack-too-deep in RateManagerV1; minimal-IR loses one branch anchor. Deterministic tests exercise configuration, authorization, fixed-rate, oracle fallback, and update paths."
+    },
+    {
+      "file": "contracts/RiskManager.sol",
+      "metric": "statements",
+      "reason": "Standard coverage cannot compile the stack-heavy RiskManager contract. Minimal-IR covers 407 of 432 statements, 96.66% of lines, 71.74% of branches, and every function without a production-only instrumentation refactor."
+    },
+    {
+      "file": "contracts/lib/ThresholdSigVerifierUtils.sol",
+      "metric": "statements",
+      "reason": "Both standard and minimal-IR coverage report one unanchored assembly-adjacent statement even though the path executes. The exact-source shard covers every line, branch, and function."
+    }
+  ]
+}
